requirements.txt
amqp==5.0.6
anyio==3.4.0
asgiref==3.4.1
attrs==20.3.0
billiard==3.6.4.0
celery==5.2.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.70.0
flower==1.0.0
h11==0.12.0
humanize==3.12.0
idna==2.10
JsonForm==0.0.2
jsonschema==3.2.0
JsonSir==0.0.2
kombu==5.2.2
orjson==3.6.4
prometheus-client==0.12.0
prompt-toolkit==3.0.23
pydantic==1.8.2
pyrsistent==0.17.3
python-dotenv==0.19.2
Python-EasyConfig==0.1.7
pytz==2021.3
PyYAML==5.4.1
redis==4.0.2
requests==2.25.1
Resource==0.2.1
six==1.15.0
sniffio==1.2.0
starlette==0.16.0
tornado==6.1
typing_extensions==4.0.0
urllib3==1.26.4
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
